質問詳細の質問内容をQuestoinDetailItemに移動させました。
